func_app/requirements.txt
azure-functions
azure-identity
azure-mgmt-network
azure-data-tables